feat(returning): support return values in the draft function (#4)
Since the user may refer to the draft in the return of a completely new object, I will use the option `strict`, which we will suggest the user enable `strict` in development mode and disable `strict` in production mode. This will ensure safe returns and also keep good performance in the production build.

If the return value is mixed drafts or `undefined`, then use `safeReturn()` wrapper.